refactor sqlwriter
