+++ pkg/util/export/etl/sql.go
@@ -16,27 +16,15 @@ package etl
 
 import (
 	"context"
-	"database/sql"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"strings"
 	"sync"
 	"time"
 
 	_ "github.com/go-sql-driver/mysql"
 	"github.com/matrixorigin/matrixone/pkg/logutil"
-	"github.com/matrixorigin/matrixone/pkg/util/export/etl/db"
+	db_holder "github.com/matrixorigin/matrixone/pkg/util/export/etl/db"
 	"github.com/matrixorigin/matrixone/pkg/util/export/table"
 	"go.uber.org/zap"
 )
-
-// MAX_CHUNK_SIZE is the maximum size of a chunk of records to be inserted in a single insert.
-const MAX_CHUNK_SIZE = 1024 * 1024 * 4
-
-const MAX_INSERT_TIME = 3 * time.Second
-
-var _ SqlWriter = (*DefaultSqlWriter)(nil)
 
 // DefaultSqlWriter SqlWriter is a writer that writes data to a SQL database.
 type DefaultSqlWriter struct {
@@ -53,11 +41,6 @@ func NewSqlWriter(ctx context.Context, tbl *table.Table, csv *CSVWriter) *Defaul
 		csvWriter: csv,
 		tbl:       tbl,
 	}
-}
-
-type SqlWriter interface {
-	table.RowWriter
-	WriteRowRecords(records [][]string, tbl *table.Table, is_merge bool) (int, error)
 }
 
 func (sw *DefaultSqlWriter) GetContent() string {
@@ -81,7 +64,7 @@ func (sw *DefaultSqlWriter) flushBuffer(force bool) (int, error) {
 	var err error
 	var cnt int
 
-	cnt, err = sw.WriteRowRecords(sw.buffer, sw.tbl, false)
+	cnt, err = db_holder.WriteRowRecords(sw.buffer, sw.tbl)
 
 	if err != nil {
 		sw.dumpBufferToCSV()
@@ -111,106 +94,4 @@ func (sw *DefaultSqlWriter) FlushAndClose() (int, error) {
 	sw.tbl = nil
 	sw.csvWriter = nil
 	return cnt, err
-}
-
-func bulkInsert(ctx context.Context, done chan error, sqlDb *sql.DB, records [][]string, tbl *table.Table, maxLen int) {
-	if len(records) == 0 {
-		done <- nil
-		return
-	}
-
-	baseStr := fmt.Sprintf("INSERT INTO `%s`.`%s` VALUES ", tbl.Database, tbl.Table)
-
-	sb := strings.Builder{}
-	defer sb.Reset()
-
-	tx, err := sqlDb.Begin()
-	if err != nil {
-		done <- err
-		return
-	}
-
-	for idx, row := range records {
-		if len(row) == 0 {
-			continue
-		}
-
-		sb.WriteString("(")
-		for i, field := range row {
-			if i != 0 {
-				sb.WriteString(",")
-			}
-			if tbl.Columns[i].ColType == table.TJson {
-				var js interface{}
-				escapedJSON, _ := json.Marshal(js)
-				_ = json.Unmarshal([]byte(field), &js)
-				sb.WriteString(fmt.Sprintf("'%s'", strings.ReplaceAll(strings.ReplaceAll(string(escapedJSON), "\\", "\\\\"), "'", "\\'")))
-			} else {
-				escapedStr := strings.ReplaceAll(strings.ReplaceAll(field, "\\", "\\\\'"), "'", "\\'")
-				if tbl.Columns[i].ColType == table.TVarchar && tbl.Columns[i].Scale < len(escapedStr) {
-					sb.WriteString(fmt.Sprintf("'%s'", escapedStr[:tbl.Columns[i].Scale-1]))
-				} else {
-					sb.WriteString(fmt.Sprintf("'%s'", escapedStr))
-				}
-			}
-		}
-		sb.WriteString(")")
-
-		if sb.Len() >= maxLen || idx == len(records)-1 {
-			stmt := baseStr + sb.String() + ";"
-			_, err := tx.ExecContext(ctx, stmt)
-			if err != nil {
-				tx.Rollback()
-				sb.Reset()
-				done <- err
-				return
-			}
-			select {
-			case <-ctx.Done():
-				// If context deadline is exceeded, rollback the transaction
-				tx.Rollback()
-				done <- errors.New("execution cancelled: context deadline exceeded")
-				return
-			default:
-			}
-			sb.Reset()
-		} else {
-			sb.WriteString(",")
-		}
-	}
-	if err := tx.Commit(); err != nil {
-		done <- err
-		return
-	}
-	done <- nil
-}
-
-func (sw *DefaultSqlWriter) WriteRowRecords(records [][]string, tbl *table.Table, is_merge bool) (int, error) {
-	if len(records) == 0 {
-		return 0, nil
-	}
-	var err error
-	dbConn, err := db_holder.InitOrRefreshDBConn(false)
-	if err != nil {
-		logutil.Error("sqlWriter db init failed", zap.Error(err))
-		return 0, err
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), MAX_INSERT_TIME)
-	defer cancel()
-
-	done := make(chan error)
-	go bulkInsert(ctx, done, dbConn, records, tbl, MAX_CHUNK_SIZE)
-
-	select {
-	case err := <-done:
-		if err != nil {
-			return 0, err
-		} else {
-			logutil.Debug("sqlWriter WriteRowRecords finished", zap.Int("cnt", len(records)))
-			return len(records), nil
-		}
-	case <-ctx.Done():
-		logutil.Warn("sqlWriter WriteRowRecords cancelled", zap.Error(ctx.Err()))
-		return 0, ctx.Err()
-	}
 }
